Rename models-nexus.ts to nexus-models.ts
