+++ packages/app/src/vis-packs/core/complex/MappedComplexLineVis.tsx
@@ -1,6 +1,6 @@
 import type { H5WebComplex } from '@h5web/shared/hdf5-models';
-import type { AxisMapping } from '@h5web/shared/models-nexus';
 import type { NumArray } from '@h5web/shared/models-vis';
+import type { AxisMapping } from '@h5web/shared/nexus-models';
 import { useMemo } from 'react';
 import { createPortal } from 'react-dom';
