+++ packages/app/src/vis-packs/core/complex/MappedComplexVis.tsx
@@ -5,8 +5,8 @@ import {
   useVisDomain,
 } from '@h5web/lib';
 import type { H5WebComplex } from '@h5web/shared/hdf5-models';
-import type { AxisMapping } from '@h5web/shared/models-nexus';
 import type { NumArray } from '@h5web/shared/models-vis';
+import type { AxisMapping } from '@h5web/shared/nexus-models';
 import { useMemo } from 'react';
 import { createPortal } from 'react-dom';
