+++ packages/app/src/vis-packs/core/heatmap/MappedHeatmapVis.tsx
@@ -5,7 +5,7 @@ import type {
   Dataset,
   NumericLikeType,
 } from '@h5web/shared/hdf5-models';
-import type { AxisMapping } from '@h5web/shared/models-nexus';
+import type { AxisMapping } from '@h5web/shared/nexus-models';
 import type { NumArray } from '@h5web/shared/models-vis';
 import { createPortal } from 'react-dom';
