+++ packages/app/src/vis-packs/core/line/MappedLineVis.tsx
@@ -5,8 +5,8 @@ import type {
   Dataset,
   NumericLikeType,
 } from '@h5web/shared/hdf5-models';
-import type { AxisMapping } from '@h5web/shared/models-nexus';
 import type { NumArray } from '@h5web/shared/models-vis';
+import type { AxisMapping } from '@h5web/shared/nexus-models';
 import { createPortal } from 'react-dom';
 
 import type { DimensionMapping } from '../../../dimension-mapper/models';
