+++ packages/app/src/vis-packs/core/rgb/MappedRgbVis.tsx
@@ -1,6 +1,6 @@
 import { RgbVis } from '@h5web/lib';
 import type { NumArrayDataset } from '@h5web/shared/hdf5-models';
-import type { AxisMapping } from '@h5web/shared/models-nexus';
+import type { AxisMapping } from '@h5web/shared/nexus-models';
 import type { NumArray } from '@h5web/shared/models-vis';
 import type { TypedArray } from 'ndarray';
 import { createPortal } from 'react-dom';
