+++ packages/shared/src/mock/metadata-utils.ts
@@ -22,7 +22,7 @@ import type {
   UnresolvedEntity,
 } from '../hdf5-models';
 import { DTypeClass, Endianness, EntityKind } from '../hdf5-models';
-import type { SilxStyle } from '../models-nexus';
+import type { SilxStyle } from '../nexus-models';
 import { buildEntityPath } from '../utils';
 import type { MockAttribute, MockDataset, MockValueId } from './models';
 import { mockValues } from './values';
